requirements.txt
aiogram==3.*
python-dotenv==0.21.0
aiosqlite
greenlet